+++ rpc/pb/wrappers/pb-marshalling.go
@@ -1,9 +1,9 @@
-// Copyright 2014 ISRG.  All rights reserved
+// Copyright 2016 ISRG.  All rights reserved
 // This Source Code Form is subject to the terms of the Mozilla Public
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-package rpc
+package wrappers
 
 import (
 	"net"
