+++ rpc/rpc-wrappers.go
@@ -16,13 +16,11 @@ import (
 	"github.com/cactus/go-statsd-client/statsd"
 	jose "github.com/square/go-jose"
 	"golang.org/x/net/context"
-	"google.golang.org/grpc"
 
 	"github.com/letsencrypt/boulder/cmd"
 	"github.com/letsencrypt/boulder/core"
 	blog "github.com/letsencrypt/boulder/log"
 	"github.com/letsencrypt/boulder/probs"
-	vaPB "github.com/letsencrypt/boulder/rpc/pb/va"
 )
 
 // This file defines RPC wrappers around the ${ROLE}Impl classes,
@@ -556,88 +554,6 @@ func (rac RegistrationAuthorityClient) OnValidationUpdate(ctx context.Context, a
 
 	_, err = rac.rpc.DispatchSync(MethodOnValidationUpdate, data)
 	return
-}
-
-type ValidationAuthorityGRPCServer struct {
-	impl core.ValidationAuthority
-}
-
-func (s *ValidationAuthorityGRPCServer) PerformValidation(ctx context.Context, in *vaPB.PerformValidationRequest) (*vaPB.ValidationResult, error) {
-	domain, challenge, authz, err := performValidationReqToArgs(in)
-	if err != nil {
-		return nil, err
-	}
-	records, err := s.impl.PerformValidation(ctx, domain, challenge, authz)
-	prob, ok := err.(*probs.ProblemDetails)
-	if !ok {
-		return nil, err
-	}
-	return validationResultToPB(records, prob)
-}
-
-func (s *ValidationAuthorityGRPCServer) IsSafeDomain(ctx context.Context, in *vaPB.IsSafeDomainRequest) (*vaPB.IsDomainSafe, error) {
-	if in == nil {
-		return nil, ErrMissingParameters
-	}
-	if in.Domain == nil {
-		return nil, ErrMissingParameters
-	}
-
-	resp, err := s.impl.IsSafeDomain(ctx, *in.Domain)
-	if err != nil {
-		return nil, err
-	}
-	return &vaPB.IsDomainSafe{Valid: &resp}, nil
-}
-
-func RegisterValidationAuthorityGRPCServer(s *grpc.Server, impl core.ValidationAuthority) error {
-	rpcSrv := &ValidationAuthorityGRPCServer{impl}
-	vaPB.RegisterVAServer(s, rpcSrv)
-	return nil
-}
-
-type ValidationAuthorityGRPCClient struct {
-	gc vaPB.VAClient
-}
-
-func NewValidationAuthorityGRPCClient(cc *grpc.ClientConn) core.ValidationAuthority {
-	return &ValidationAuthorityGRPCClient{vaPB.NewVAClient(cc)}
-}
-
-func (vac ValidationAuthorityGRPCClient) UpdateValidations(ctx context.Context, authz core.Authorization, index int) error {
-	panic("UpdateValidations should not be called on VA GRPC client")
-}
-
-// PerformValidation has the VA revalidate the specified challenge and returns
-// the updated Challenge object.
-func (vac ValidationAuthorityGRPCClient) PerformValidation(ctx context.Context, domain string, challenge core.Challenge, authz core.Authorization) ([]core.ValidationRecord, error) {
-	req, err := argsToPerformValidationRequest(domain, challenge, authz)
-	if err != nil {
-		return nil, err
-	}
-	gRecords, err := vac.gc.PerformValidation(ctx, req)
-	if err != nil {
-		return nil, err
-	}
-	records, prob, err := pbToValidationResult(gRecords)
-	if err != nil {
-		return nil, err
-	}
-
-	return records, prob
-}
-
-// IsSafeDomain returns true if the domain given is determined to be safe by an
-// third-party safe browsing API.
-func (vac ValidationAuthorityGRPCClient) IsSafeDomain(ctx context.Context, domain string) (isSafe bool, err error) {
-	valid, err := vac.gc.IsSafeDomain(ctx, &vaPB.IsSafeDomainRequest{Domain: &domain})
-	if err != nil {
-		return false, err
-	}
-	if valid == nil || (*valid).Valid == nil {
-		return false, ErrMissingParameters
-	}
-	return *(*valid).Valid, nil
 }
 
 // NewValidationAuthorityServer constructs an RPC server
